refactor(frontend): move the VM::exec* families into ir/vm/
The plan asks for `execute()` to shrink to a dispatch table and for VM.cpp to
land near 350 lines. The previous commit did the first half but left all 18
exec* bodies in VM.cpp, which actually grew from 1775 to 1855 lines. This
finishes the job by relocating the bodies:

  vm/VMValue.cpp      valueToString / typeNameOf / cloneValue / valueToBool
  vm/VMExecArith.cpp  applyArithmetic / applyUnary / applyComparison
                      execArithmetic / execComparison / execLogic
  vm/VMExecStack.cpp  push / pop / storeVariable / pushFrame
                      execPushConst / execStackManip / execOptional
                      execLocalSlot / execVariable
  vm/VMExecObject.cpp execFieldAccess / execArray / execClosure
                      execInstanceof / execObjectCreate
  vm/VMExecCall.cpp   execMethodDispatch / execNativeCall
                      execFunctionCall / execHostCall

VM.cpp keeps run(), execute(), failBudget(), execBranch() (22 lines of control
flow that belong with the main loop), callFunctionRef() and the two class
helpers: 1855 -> 296 lines, against the plan's ~350 target.

Two details worth calling out:

tlsBudget is a mutable thread_local in VM.cpp's anonymous namespace. It cannot
be copied into the new files the way an ordinary helper can -- every copy would
be a distinct object and the budget would stop propagating. It stays in VM.cpp,
where the only user (execute()) lives. splitStaticMemberName and
valueClassNameOf do move, each to the single vm/ file that calls it, and each
in an anonymous namespace so it stays out of the unity build's symbol table.

VM.h renames the globals constructor parameter to globalsTable. The old name
shadowed the member of the same name, and because the warning lives in a header
the split multiplied it from 9 to 24 instances across the build. Renaming a
parameter cannot affect callers.

Verification: 429 tests pass, same 5 pre-existing failures; bytecode dump
byte-identical. -Wall -Wextra -Wshadow over every added and modified file now
reports 3 diagnostics, all of them present in the base commit.
